// Job_Tracker/src/models/JobInfo.ts
export class JobInfo{
    userId: Number;
    companyName: String;
    jobTitle: String;
    jobDescription: String;
    jobStatus: String;
    interviewDate1: Date;
    interviewDate2: Date;
    appliedDate: Date;
    offerDate: Date;
    jobResume:String;
    jobCoverLetter: String;
    salary: Number;
    location: String;

}